move settings definition to index
